+++ src/main/java/com/example/shoppingmallserver/repository/MileageRepository.java
@@ -1,0 +1,8 @@
+package com.example.shoppingmallserver.repository;
+
+import com.example.shoppingmallserver.entity.Mileage;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MileageRepository extends JpaRepository<Mileage, Long> {
+}
