+++ src/main/java/com/example/shoppingmallserver/service/UserServiceImpl.java
@@ -18,83 +18,41 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
+    // 사용자 정보 조회
     @Override
     public User getUserById(Long id) {
         return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
     }
 
-    @Override
-    public List<Cart> getCartItemByUserId(Long user_id) {
-        return null;
-    }
+    // ==========================관리자===========================
 
-    @Override
-    public Cart addCartItem(Long user_id, AddCartItemDto addCartItemDto) {
-        return null;
-    }
-
-    @Override
-    public void deleteCartItem(Long user_id) {
-
-    }
-
-    @Override
-    public List<Cart> purchaseCartItem(Long id, PurchaseCartItemDto purchaseCartItemDto) {
-        return null;
-    }
-
-    @Override
-    public Mileage requestMileageRecharge(Long user_id, ChargeMileageDto chargeMileageDto) {
-        return null;
-    }
-
-    @Override
-    public Wishlist addWishlistItem(Long user_id, AddWishlistItemDto addWishlistItemDto) {
-        return null;
-    }
-
-    @Override
-    public List<Wishlist> getWishlistItemList(Long user_id) {
-        return null;
-    }
-
-    @Override
-    public void deleteWishlistItem(Long id) {
-
-    }
-
+    // 관리자의 사용자 정보 조회
     @Override
     public User getAdminUserById(Long user_id) {
         return null;
     }
 
+    // 관리자의 사용자 정보 리스트 조회
     @Override
     public List<User> getUserList(Optional<String> keyword) {
         return null;
     }
 
+    // 관리자의 사용자 정보 수정
     @Override
     public User updateUser(Long user_id, UpdateUserDto updateUserDto) {
         return null;
     }
 
+    // 관리자의 사용자 정보 등록
     @Override
     public User createUser(CreateUserDto userCreateDto) {
         return null;
     }
 
+    // 관리자의 사용자 정보 삭제
     @Override
     public void deleteUser(Long user_id) {
 
-    }
-
-    @Override
-    public List<User> getMileageList(Optional<String> keyword) {
-        return null;
-    }
-
-    @Override
-    public Mileage updateMileageStatus(Long user_id, UpdateMileageStatusDto updateMileageStatusDto) {
-        return null;
     }
 }
